Reset code to master
To ease merging of the chumsky-parser-2 branch, where we re-tried the
implementation of a chumsky based parser and which works way better (in
terms of progress) than this branch, we reset the state of this branch
to master and merge in the chumsky-parser-2 branch with the next commit.

I feel like keeping this history is just the right thing to do, this is
why I do not simply delete this branch.

Signed-off-by: Matthias Beyer <mail@beyermatthias.de>