Merge Token::Ctrl, Token::Op and Token::Arrow
This merges the tokens into one: Token::Op.

It is less complex to parse this way, especially because we have to
order the parsers in the right order to parse long tokens first and
shorters after.
This solves also the bug that an "->" token was not parsed correctly
(because I've put it in the ident() parser where it absolutely does not
belong).

Signed-off-by: Matthias Beyer <mail@beyermatthias.de>