Reintroduce DefRhs type
This introduces the DefRhs type, which represents the RHS of a
definition, which can be

* A Boolean
* An integer
* A float
* A string literal
* A list of expressions
* A Unary
* A Binary
* A let-in block
* An if-else block
* A Func definition

All of these must be compiletime-evaluatable though, so if we do

    foo = if func
      then 1
      else 2

or

    foo = ! (func 1)

or

    foo = 1 + (func 1)

or

    foo = let
        bar = 1;
        baz = 2;
        in func 1 + 2;

then `func` must also be (in all cases) something that is compiletime
evaluatable.

Besides that can we define `foo` also to be curried from other
functions:

    add = (a: I8, b: I8) -> a + b;

    addOne = add 1

So here `addOne` is now `(b: I8) -> 1 + b;`.

Signed-off-by: Matthias Beyer <mail@beyermatthias.de>